Split types into matching files
