+++ src/MediatR/IAsyncNotification.cs
@@ -1,0 +1,7 @@
+namespace MediatR
+{
+    /// <summary>
+    /// Marker interface to represent an asynchronous notification
+    /// </summary>
+    public interface IAsyncNotification { }
+}
