+++ src/MediatR/ICancellableAsyncNotification.cs
@@ -1,0 +1,7 @@
+namespace MediatR
+{
+    /// <summary>
+    /// Marker interface to represent a cancellable, asynchronous notification
+    /// </summary>
+    public interface ICancellableAsyncNotification { }
+}
