+++ src/MediatR/INotification.cs
@@ -1,0 +1,7 @@
+namespace MediatR
+{
+    /// <summary>
+    /// Marker interface to represent a notification
+    /// </summary>
+    public interface INotification { }
+}
