+++ src/MediatR/IRequest.cs
@@ -6,45 +6,8 @@
     public interface IRequest : IRequest<Unit> { }
 
     /// <summary>
-    /// Marker interface to represent an asynchronous request with a void response
-    /// </summary>
-    public interface IAsyncRequest : IAsyncRequest<Unit> { }
-
-    /// <summary>
-    /// Marker interface to represent a cancellable, asynchronous request with a void response
-    /// </summary>
-    public interface ICancellableAsyncRequest : ICancellableAsyncRequest<Unit> { }
-
-    /// <summary>
     /// Marker interface to represent a request with a response
     /// </summary>
     /// <typeparam name="TResponse">Response type</typeparam>
     public interface IRequest<out TResponse> { }
-
-    /// <summary>
-    /// Marker interface to represent an asynchronous request with a response
-    /// </summary>
-    /// <typeparam name="TResponse">Response type</typeparam>
-    public interface IAsyncRequest<out TResponse> { }
-
-    /// <summary>
-    /// Marker interface to represent a cancellable, asynchronous request with a response
-    /// </summary>
-    /// <typeparam name="TResponse">Response type</typeparam>
-    public interface ICancellableAsyncRequest<out TResponse> { }
-
-    /// <summary>
-    /// Marker interface to represent a notification
-    /// </summary>
-    public interface INotification { }
-
-    /// <summary>
-    /// Marker interface to represent an asynchronous notification
-    /// </summary>
-    public interface IAsyncNotification { }
-
-    /// <summary>
-    /// Marker interface to represent a cancellable, asynchronous notification
-    /// </summary>
-    public interface ICancellableAsyncNotification { }
 }
