+++ src/MediatR/IRequestHandler.cs
@@ -1,0 +1,18 @@
+namespace MediatR
+{
+    /// <summary>
+    /// Defines a handler for a request
+    /// </summary>
+    /// <typeparam name="TRequest">The type of request being handled</typeparam>
+    /// <typeparam name="TResponse">The type of response from the handler</typeparam>
+    public interface IRequestHandler<in TRequest, out TResponse>
+        where TRequest : IRequest<TResponse>
+    {
+        /// <summary>
+        /// Handles a request
+        /// </summary>
+        /// <param name="message">The request message</param>
+        /// <returns>Response from the request</returns>
+        TResponse Handle(TRequest message);
+    }
+}
